embedding_service/requirements.txt
sentence-transformers
fastapi
pyyaml
uvicorn
pydantic
